add new updated and differently structured tests with better coverage
