+++ tests/test_core_fileproc.py
